[glow][Executor] Move ExecutionState to private header
