+++ lib/Runtime/Executor/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_library(Executor
+              ExecutionState.cpp
               ThreadPoolExecutor.cpp)
 
 target_link_libraries(Executor
